+++ tests/Serilog.Sinks.RabbitMQ.Tests/RabbitMQClientConfigurationTests.cs
@@ -36,7 +36,7 @@ public class RabbitMQClientConfigurationTests
         };
 
         // Act
-        var copy = new RabbitMQClientConfiguration().From(original);
+        var copy = original.Clone();
 
         // Assert
         copy.ShouldBeEquivalentTo(original);